LPD-19822 DuplicateQueryStringsDetector is an internal interface, it should not be registered as osgi service
