+++ modules/dxp/apps/portal-search-tuning/portal-search-tuning-rankings-web/src/main/java/com/liferay/portal/search/tuning/rankings/web/internal/portlet/action/EditRankingMVCActionCommand.java
@@ -22,7 +22,9 @@ import com.liferay.portal.kernel.util.Portal;
 import com.liferay.portal.kernel.util.StringUtil;
 import com.liferay.portal.kernel.util.Validator;
 import com.liferay.portal.kernel.util.WebKeys;
+import com.liferay.portal.search.engine.adapter.SearchEngineAdapter;
 import com.liferay.portal.search.index.IndexNameBuilder;
+import com.liferay.portal.search.query.Queries;
 import com.liferay.portal.search.tuning.rankings.constants.ResultRankingsConstants;
 import com.liferay.portal.search.tuning.rankings.helper.RankingHelper;
 import com.liferay.portal.search.tuning.rankings.index.Ranking;
@@ -54,6 +56,7 @@ import javax.portlet.ActionResponse;
 import javax.portlet.PortletConfig;
 import javax.portlet.PortletRequest;
 
+import org.osgi.service.component.annotations.Activate;
 import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.Reference;
 
@@ -98,6 +101,12 @@ public class EditRankingMVCActionCommand extends BaseMVCActionCommand {
 		}
 	}
 
+	@Activate
+	protected void activate() {
+		_duplicateQueryStringsDetector = new DuplicateQueryStringsDetector(
+			_queries, _searchEngineAdapter);
+	}
+
 	protected String getIndexName(ActionRequest actionRequest) {
 		return indexNameBuilder.getIndexName(
 			portal.getCompanyId(actionRequest));
@@ -106,9 +115,6 @@ public class EditRankingMVCActionCommand extends BaseMVCActionCommand {
 	protected RankingIndexName getRankingIndexName() {
 		return rankingIndexNameBuilder.getRankingIndexName(_companyId);
 	}
-
-	@Reference
-	protected DuplicateQueryStringsDetector duplicateQueryStringsDetector;
 
 	@Reference
 	protected IndexNameBuilder indexNameBuilder;
@@ -264,7 +270,7 @@ public class EditRankingMVCActionCommand extends BaseMVCActionCommand {
 		Ranking ranking, Collection<String> queryStrings) {
 
 		List<String> duplicateQueryStrings =
-			duplicateQueryStringsDetector.detect(
+			_duplicateQueryStringsDetector.detect(
 				new Criteria.Builder(
 				).groupExternalReferenceCode(
 					ranking.getGroupExternalReferenceCode()
@@ -666,6 +672,13 @@ public class EditRankingMVCActionCommand extends BaseMVCActionCommand {
 	private static final String _UPDATE_SPECIAL = StringPool.GREATER_THAN;
 
 	private long _companyId;
+	private DuplicateQueryStringsDetector _duplicateQueryStringsDetector;
+
+	@Reference
+	private Queries _queries;
+
+	@Reference
+	private SearchEngineAdapter _searchEngineAdapter;
 
 	private class EditRankingMVCActionRequest {
